backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-genai==0.3.0
python-multipart==0.0.9
pydantic==2.6.4
motor==3.3.2
pymongo==4.6.2
python-dotenv==1.0.1
